requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.10.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
faker==33.1.0
bcrypt==4.2.1
PyJWT==2.10.1
